feat(#84): warn on missing_docs, handlers for better consistency
