+++ server/src/handlers/home.rs
@@ -27,10 +27,7 @@ use log::debug;
 use crate::ServerConfig;
 use openapi::models::MetaRoot200Response;
 
-// @todo #79:45min Make mod/fn/struct documentation required.
-//  Let's configure cargo in such way, so documentation will be required at all
-//  levels. When such rule will be configured, please fix all warnings/errors
-//  and integrate changes into master branch.
+/// Home handler.
 pub async fn home(State(config): State<ServerConfig>) -> impl IntoResponse {
     let response = Json(
         MetaRoot200Response::new(
@@ -191,7 +188,7 @@ pub async fn home(State(config): State<ServerConfig>) -> impl IntoResponse {
 
 #[cfg(test)]
 mod tests {
-    use crate::routes::home::home;
+    use crate::handlers::home::home;
     use crate::ServerConfig;
     use anyhow::Result;
     use axum::body::to_bytes;
