+++ server/src/handlers/mod.rs
@@ -19,15 +19,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-use axum::http::StatusCode;
-use axum::Json;
-
-use crate::objects::user::User;
-
-pub async fn register_user(Json(payload): Json<User>) -> Result<StatusCode, String> {
-    let user = User::new(payload.username.clone());
-    match user.save().await {
-        Ok(_) => Ok(StatusCode::CREATED),
-        Err(e) => Err(format!("Can't register {}: {}", payload.username, e)),
-    }
-}
+/// Home handler.
+pub mod home;
+/// Handler for internal user registration.
+pub mod register_user;
